requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
braceexpand==0.1.7
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py @ git+https://github.com/Rapptz/discord.py@a5d03d4a1e6f4314518f6f55d8b7130ab3f4336e
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
import-expression==1.1.4
Jinja2==3.1.2
jishaku==2.5.2
MarkupSafe==2.1.3
multidict==6.0.4
nodeenv==1.8.0
prisma==0.11.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
tomlkit==0.12.2
typing_extensions==4.8.0
yarl==1.9.2